+++ src/lib/Core/CHANGELOG.md
@@ -1,6 +1,7 @@
 ﻿# 3.1.5
 * Renamed Matrix4.empty and Vector4.create to Matrix4/Vector4.zero.
 * Added more linear algebra functions.
+* Removed ObjectVec3 from linear algebra.
 
 # 3.1.4
 * Added "allAtoms" query.